# docs/swagger.yaml
definitions:
...
  request.CreateTokenRequest:
    properties:
      expires_at:
        description: optional RFC3339
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  response.CreateTokenResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      is_revoked:
        type: boolean
      last_used_at:
        type: string
      name:
        type: string
      raw_token:
        description: only returned once on create
        type: string
      token_prefix:
        type: string
    type: object
...
  response.TokenResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      is_revoked:
        type: boolean
      last_used_at:
        type: string
      name:
        type: string
      token_prefix:
        type: string
    type: object
  utils.ErrorResponseModel:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
  utils.SuccessResponseModel:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api-tokens:
    get:
      description: Get all API tokens for the authenticated user (admins see all)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.TokenResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: List API tokens
      tags:
      - API Tokens
    post:
      consumes:
      - application/json
      description: Generate a new API token. Raw token is only returned once.
      parameters:
      - description: Token details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CreateTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create API token
      tags:
      - API Tokens
  /api-tokens/{id}:
    delete:
      description: Revoke an API token by ID. Immediate loss of access.
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Revoke API token
      tags:
      - API Tokens
...
  /api/backup/download:
    get:
      consumes:
      - application/json
      description: Download a JSON backup file as an attachment. Requires admin privileges.
      produces:
      - application/json
      responses:
        "200":
          description: Backup JSON file download
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: '[Admin] Download backup as file'
      tags:
      - Backup
  /api/backup/export:
    get:
      consumes:
      - application/json
      description: Export all configuration data as a JSON backup. Requires admin
        privileges.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: '[Admin] Export backup data'
      tags:
      - Backup
  /api/backup/history:
    get:
      consumes:
      - application/json
      description: Retrieve backup/import history records. Requires admin privileges.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: '[Admin] List backup history'
      tags:
      - Backup
  /api/backup/history/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a backup/import history record by ID. Requires admin privileges.
      parameters:
      - description: Backup record ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: '[Admin] Delete backup history record'
      tags:
      - Backup
  /api/backup/import:
    post:
      consumes:
      - multipart/form-data
      description: Import configuration from a JSON backup file upload (multipart).
        Requires admin privileges.
      parameters:
      - description: Backup JSON file to import
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponseModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: '[Admin] Import backup data'
      tags:
      - Backup
...
swagger: "2.0"